fix test lint

+++ pacific-architecture/data/src/main/java/com/pacific/data/db/UserDao.kt
@@ -8,5 +8,5 @@ import com.pacific.data.db.entities.DbUser
 interface UserDao : AppDao<DbUser> {
 
     @Query("SELECT * FROM user WHERE _id =:id")
-    fun get(id: Int): DbUser?
+    fun get(id: Long): DbUser?
 }
